Added handling for txt/html responses

+++ src/pypiwizard/pypi_dl.py
@@ -5,22 +5,30 @@ Created on 2025-02-04T09:19:05-05:00
 
 @author: nate
 """
+import argparse
+import asyncio
+import atexit
+import datetime
+import json
+import logging
 import os
 import re
-import sys
-import json
-import argparse
-import subprocess as sp
-import datetime
-import atexit
-
-import asyncio
-import aiohttp
-import feud
 import shlex
+import subprocess as sp
+import sys
+from urllib.parse import urlparse
+
+import aiohttp
+import bs4
+import patoolib
+import rich_click as click
+from dataclass_click import dataclass_click
+from loguru import logger
+
 import pypiwizard as ppw
 
-logger = ppw.util.get_logger(__name__)
+patoolib.log.logger.setLevel(logging.WARN)
+
 
 def subproc(cmd):
     shlex.split(cmd)
@@ -31,7 +39,6 @@ outpath = os.path.join(os.environ['HOME'], "pypiwizard_sdists")
 index_file = os.path.join(outpath, 'simple.json')
 os.makedirs(outpath, exist_ok=True)
 
-
 async def download_index():
     url = "https://pypi.org/simple/?format=application/vnd.pypi.simple.v1+json"
     async with aiohttp.ClientSession() as sesh:
@@ -41,13 +48,21 @@ async def download_index():
             fp.write(json.dumps(data, indent=2))
         return data
 
-async def dl_tarball(sesh, tb):
-    url = tb['url']
-    resp = await sesh.request('GET', url=url)
-    tb_path = os.path.join(outpath, tb['filename'])
+async def dl_tarball(sesh, tb_url, filename):
+
+    if os.path.exists(filename):
+        logger.info(f'File {filename} already exists')
+        return
+
+    resp = await sesh.request('GET', url=tb_url)
+    tb_path = os.path.join(outpath, filename)
     bits = await resp.content.read()
     with open(tb_path, "w+b") as fp:
         fp.write(bits)
+
+    new_path = patoolib.extract_archive(tb_path, outdir=outpath)
+    os.unlink(tb_path)
+    return
 
     cmd = f"""
 pushd '{outpath}' \
@@ -58,14 +73,45 @@ pushd '{outpath}' \
     subproc(cmd)
     return
 
+def get_wheel(text):
+    soup = bs4.BeautifulSoup(text, 'html.parser')
+    last_href = ""
+    wheels = []
+    tarballs = []
+    for link in soup.find_all('a'):
+        href = link.attrs.get('href')
+        if not href:
+            breakpoint()
+            continue
+
+        parsed = urlparse(href)
+        path = parsed.path
+        if path.endswith('.tar.gz'):
+            tarballs.append(href)
+            continue
+        elif path.endswith('.whl'):
+            wheels.append(href)
+            continue
+
+        breakpoint()
+
+    return wheels, tarballs
+
+
 async def retrieve_project(sem0, sesh, proj_name):
     async with sem0:
         url = f"https://pypi.org/simple/{proj_name}/?format=application/vnd.pypi.simple.v1+json"
         resp = await sesh.request("GET", url)
         txt = await resp.text()
         if resp.headers.get('Content-Type') == 'text/html':
-            return
-            # TODO: handle this better
+
+            wheels, tarballs = get_wheel(txt)
+            if len(tarballs) == 0:
+                logger.info(f'No tarballs found for {proj_name}...')
+            url0 = tarballs[-1]
+            fname0 = os.path.basename(urlparse(url0).path)
+            await dl_tarball(sesh, url0, fname0)
+            return wheels, tarballs
 
         data = json.loads(txt)
         tarballs = []
@@ -77,12 +123,13 @@ async def retrieve_project(sem0, sesh, proj_name):
             tarballs.append(item)
         ##############################################################
         if len(tarballs) == 0:
-            return
+            logger.info(f'No sdist found for {proj_name}')
+            breakpoint()
+            return      
 
         tarballs = sorted(tarballs, key=lambda item: item['upload-time'], reverse=True)
         tb = tarballs[0]
-        await dl_tarball(sesh, tb)
-
+        await dl_tarball(sesh, tb['url'], tb['filename'])
         return wheels, tarballs
 
 
@@ -119,7 +166,6 @@ async def do_search(regex: str, index):
             proj_name = proj['name']
             if not re.match(regex, proj_name.lower()):
                 continue
-
             task = asyncio.create_task(get_info(sem0, sesh, proj_name))
             tasks.append(task)
 
@@ -139,68 +185,72 @@ def load_or_fetch_index(update_index:bool):
 
     return data
 
-class SimpleDl(feud.Group):
-    """Download sdists from PyPi."""
-    def dl(regex: str, *, update_index: bool = False, download: bool = False):
-        """Download sdists via regex."""
-        data = load_or_fetch_index(update_index)
-        all_projects = data['projects']
-        matched = {}
-        for i, proj in enumerate(all_projects):
-            proj_name = proj['name']
-            if not re.match(regex, proj_name):
-                continue
-            matched[proj_name] = proj
-            logger.info(f'{i:8}: {proj_name}')
-        if not download:
-            return
-        asyncio.run(get_projects(matched))
 
-    def search(regex: str, *, update_index: bool = False):
-        """Search for packages by regex."""
-        data = load_or_fetch_index(update_index)
-        result = asyncio.run(do_search(regex, data))
-        for item in result:
-            if not 'info' in item:
-                continue
-            info = item['info']
-            name = info['name']
-            summary = info['summary']
-            print(f"{name:50}: {summary}")
+@click.command()
+@dataclass_click(ppw.types.SearchArgs)
+def search(args: ppw.types.SearchArgs):
+    """Search for packages by regex."""
+    data = load_or_fetch_index(args.update_index)
+    result = asyncio.run(do_search(args.regex, data))
+    for item in result:
+        if not 'info' in item:
+            continue
+        info = item['info']
+        name = info['name']
+        summary = info['summary']
+        print(f"{name:50}: {summary}")
 
-    def scored(regex: str, *, update_index: bool = False):
-        """Retrieve the top 100 results."""
-        data = load_or_fetch_index(update_index)
-        result = asyncio.run(do_search(regex, data))
-        scores = {}
-        specs = {}
-        rev = None
-        for item in result:
-            if not 'info' in item:
-                continue
-            info = item['info']
-            name = info['name']
-            specs[name] = item
-            def get(d0, key):
-                if key in d0 and d0[key]:
-                    return len(d0[key])
-                return 0
-            # A decent heuristic popularity score
-            # Ranks aiohttp as #1 out of thousands when you search for 'aio.*'
-            rev = True
-            score1 = get(item, 'urls')
-            score2 = get(info, 'requires_dist')
-            score3 = get(info, 'classifiers')
-            score = score1+score2+score3
-            scores[name] = score
-            print(f"{name}: {score}")
 
-        result = sorted(scores.items(), key=lambda item: item[1], reverse=rev)
-        result = result[0:100]
-        for i, (item, score) in enumerate(result):
-            spec = specs[item]
-            summary = spec['info']['summary']
-            print(f"[{i:3}][{score:4}] {item} - {summary}")
+@click.command()
+@dataclass_click(ppw.types.SearchArgs)
+def search_scored(args: ppw.types.SearchArgs):
+    """Retrieve the top 100 results."""
+    data = load_or_fetch_index(args.update_index)
+    result = asyncio.run(do_search(args.regex, data))
+    scores = {}
+    specs = {}
+    rev = None
+    for item in result:
+        if not 'info' in item:
+            continue
+        info = item['info']
+        name = info['name']
+        specs[name] = item
+        def get(d0, key):
+            if key in d0 and d0[key]:
+                return len(d0[key])
+            return 0
+        # A decent heuristic popularity score
+        # Ranks aiohttp as #1 out of thousands when you search for 'aio.*'
+        rev = True
+        score1 = get(item, 'urls')
+        score2 = get(info, 'requires_dist')
+        score3 = get(info, 'classifiers')
+        score = score1+score2+score3
+        scores[name] = score
+        #print(f"{name}: {score}")
 
-if __name__ == "__main__":
-    feud.run(SimpleDl)
+    result = sorted(scores.items(), key=lambda item: item[1], reverse=rev)
+    result = result[0:100]
+    for i, (item, score) in enumerate(result):
+        spec = specs[item]
+        summary = spec['info']['summary']
+        print(f"[{i:3}][{score:4}] {item} - {summary}")
+
+
+@click.command()
+@dataclass_click(ppw.types.DlSimpleArgs)
+def dl_simple(args: ppw.types.DlSimpleArgs):
+    """Download sdists via regex."""
+    data = load_or_fetch_index(args.update_index)
+    all_projects = data['projects']
+    matched = {}
+    for i, proj in enumerate(all_projects):
+        proj_name = proj['name']
+        if not re.match(args.regex, proj_name):
+            continue
+        matched[proj_name] = proj
+        logger.info(f'{i:8}: {proj_name}')
+    if not args.download:
+        return
+    asyncio.run(get_projects(matched))
